% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{perpetual_base_score}
\alias{perpetual_base_score}
\title{Get the base score of the model}
\usage{
perpetual_base_score(model)
}
\arguments{
\item{model}{A \code{PerpetualBooster} object.}
}
\value{
Numeric, base score.
}
\description{
Get the base score of the model
}
