% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{perpetual_importance}
\alias{perpetual_importance}
\title{Get feature importance}
\usage{
perpetual_importance(model, method = "gain", normalize = TRUE)
}
\arguments{
\item{model}{A \code{PerpetualBooster} object.}

\item{method}{String, method for importance (e.g. "gain").}

\item{normalize}{Boolean.}
}
\value{
A named vector of importances.
}
\description{
Get feature importance
}
