% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{perpetual_save}
\alias{perpetual_save}
\title{Save a PerpetualBooster model}
\usage{
perpetual_save(model, path)
}
\arguments{
\item{model}{A \code{PerpetualBooster} object.}

\item{path}{String, path to save the model.}
}
\description{
Save a PerpetualBooster model
}
