% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{perpetual_to_json}
\alias{perpetual_to_json}
\title{Dump model to JSON string}
\usage{
perpetual_to_json(model)
}
\arguments{
\item{model}{A \code{PerpetualBooster} object.}
}
\value{
String, JSON representation of the model.
}
\description{
Dump model to JSON string
}
