% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{data}
\name{PerpetualBooster}
\alias{PerpetualBooster}
\title{PerpetualBooster}
\format{
An object of class \code{list} of length 2.
}
\usage{
PerpetualBooster
}
\description{
PerpetualBooster object for creating and loading models.
}
\keyword{datasets}
