% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{predict.PerpetualBooster}
\alias{predict.PerpetualBooster}
\title{Predict using a PerpetualBooster model}
\usage{
\method{predict}{PerpetualBooster}(
  object,
  newdata,
  type = c("class", "prob", "raw", "contribution", "interval"),
  method = "average",
  ...
)
}
\arguments{
\item{object}{A \code{PerpetualBooster} object.}

\item{newdata}{A matrix or data.frame of new data to predict on.}

\item{type}{Type of prediction: "class", "prob", "raw", "contribution", or "interval".}

\item{method}{Method for prediction contributions. Default "average".}

\item{...}{Additional arguments.}
}
\value{
A vector or matrix of predictions.
}
\description{
Predict using a PerpetualBooster model
}
